tg-module/requirements.txt
aiogram==3.2.0
aiohttp==3.9.1
asyncio-throttle==1.0.2
python-dotenv==1.0.0
uvicorn==0.24.0
